requirements.txt
Django==1.6
django-grappelli==2.5.0
django-tastypie==0.11.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.4.1
wsgiref==0.1.2
